Seal class Robot
I cheated the interface by creating setters for Name and Email
that are actually no-ops.  This means the consumer cannot
modify these fields even though the interface IPlayerModel
indicates they have setters, which implies that they can be
changed.  The fact that the setters for these properties in
Robot are no-ops subverts the expectations one might have
based on the interface that PlayerModel types must implement.

For this reason, Robot is sealed.  Let this be the last class
in this lineage who deviates from the standard.  The sin of
nonconformity has left Robot estranged from its virile powers,
as the Program sees to the end of Robot's generations.