Elide Program.Exceptions
NoInputException used to be the exit path from the program execution,
but this makes little sense as it would be caught by the main
program loop from deeper subroutines and throw the user out of
the program possibly unexpectedly.  Furthermore, no-input is not
conventionally a quit option in other programs, but rather a way of
submitting default input automatically, which isn't applicable to this
program.

Replacing the exit-path with a single 'q' follows convention
and makes the program's exit-path explicit and singular.