Run program as an instance
It kind of bothered me that Main() was responsible not only for
running the program, but also for initializing the _players field.
I figured that business was more relevant to the domain of a
class constructor, which I could only have if I made the whole
program run as an instance of the class.  I don't know whether
there is any meaningful difference, whether this is a pattern or
an anti-pattern, but now Program.MainLoop() has no other behavior
than to execute and run the main program loop, so that's cool.